+++ packages/components/src/FormField/DEPRACATED_FormField.css.d.ts
@@ -1,0 +1,15 @@
+export const wrapper: string;
+export const formField: string;
+export const label: string;
+export const select: string;
+export const inline: string;
+export const small: string;
+export const large: string;
+export const center: string;
+export const right: string;
+export const invalid: string;
+export const disabled: string;
+export const miniLabel: string;
+export const textareaLabel: string;
+export const maxLength: string;
+export const postfix: string;
