+++ packages/components/src/FormField/FormField.css.d.ts
@@ -1,15 +1,16 @@
 export const wrapper: string;
-export const formField: string;
-export const label: string;
-export const select: string;
-export const inline: string;
-export const small: string;
-export const large: string;
-export const center: string;
-export const right: string;
+export const textarea: string;
 export const invalid: string;
 export const disabled: string;
-export const miniLabel: string;
-export const textareaLabel: string;
+export const small: string;
+export const large: string;
+export const inline: string;
+export const center: string;
+export const right: string;
 export const maxLength: string;
+export const inputWrapper: string;
+export const input: string;
+export const miniLabel: string;
+export const select: string;
+export const label: string;
 export const postfix: string;
