--- packages/components/src/FormField/FormFieldStyles.css
@@ -1,212 +1,0 @@
-:root {
-  --field--color: var(--color-blue);
-  --field--height: var(--space-largest);
-  --field--textAlign: left;
-
-  --field--padding-top: calc(var(--space-base) - var(--space-smallest));
-  --field--padding-bottom: calc(var(--space-base) - var(--space-smallest));
-  --field--padding-left: var(--space-base);
-  --field--padding-right: var(--space-base);
-
-  --field--border-color: var(--color-border);
-  --field--background-color: var(--color-surface);
-
-  --field--placeholder-color: var(--color-greyBlue--light);
-
-  --field-base-elevation: var(--elevation-base);
-
-  /* TODO: Document these, they are weird */
-  --public-field--top-left-radius: var(--radius-base);
-  --public-field--top-right-radius: var(--radius-base);
-  --public-field--bottom-left-radius: var(--radius-base);
-  --public-field--bottom-right-radius: var(--radius-base);
-}
-
-/** 
- * Wrapper
- **/
-.wrapper {
-  display: flex;
-  width: 100%;
-  height: var(--field--height);
-  border: var(--border-base) solid var(--field--border-color);
-  border-radius: var(--radius-base);
-  border-top-right-radius: var(--public-field--top-right-radius);
-  border-bottom-right-radius: var(--public-field--bottom-right-radius);
-  border-bottom-left-radius: var(--public-field--bottom-left-radius);
-  border-top-left-radius: var(--public-field--top-left-radius);
-  color: var(--field--color);
-  font-size: var(--base-unit);
-  background-color: var(--field--background-color);
-}
-
-.wrapper * {
-  box-sizing: border-box;
-}
-
-.wrapper:focus-within {
-  --field--border-color: var(--color-focus);
-}
-
-.wrapper.textarea {
-  height: auto;
-  min-height: auto;
-}
-
-.invalid,
-.invalid:focus-within {
-  --field--border-color: var(--color-critical);
-}
-
-.disabled {
-  --field--placeholder-color: var(--color-disabled);
-  --field--background-color: var(--color-surface--background);
-  --field--border-color: var(--color-disabled--secondary);
-
-  opacity: 0.6;
-}
-
-.small {
-  --field--padding-left: calc(var(--space-base) - var(--space-smaller));
-  --field--padding-right: calc(var(--space-base) - var(--space-smaller));
-  --field--height: calc(var(--space-larger) + var(--space-smaller));
-}
-
-.large {
-  --field--padding-left: var(--space-large);
-  --field--padding-right: var(--space-large);
-  --field--height: calc(var(--space-extravagant));
-}
-
-.inline {
-  display: inline-flex;
-  width: auto;
-}
-
-.center {
-  --field--textAlign: center;
-}
-
-.right {
-  --field--textAlign: right;
-}
-
-/**
- * The custom property --formField-maxLength is getting defined inside the
- * component.
- */
-.maxLength {
-  width: calc(
-    (var(--base-unit) * var(--formField-maxLength)) + var(--field--padding-left) +
-      var(--field--padding-right)
-  );
-}
-
-/**
- * Input Wrapper
- **/
-.inputWrapper {
-  display: flex;
-  position: relative;
-  flex: 1;
-}
-
-/**
- * Input
- **/
-.input {
-  position: relative;
-  z-index: var(--field-base-elevation);
-  width: 100%;
-  padding-top: var(--field--padding-top);
-  padding-bottom: var(--field--padding-bottom);
-  padding-left: var(--field--padding-left);
-  padding-right: var(--field--padding-right);
-  border: none;
-  border-radius: 0;
-  color: inherit;
-  font-family: inherit;
-  font-size: inherit;
-  line-height: calc(var(--base-unit) * 1.25);
-  text-align: var(--field--textAlign);
-  background: transparent;
-  appearance: none;
-}
-
-.input:focus {
-  outline: none;
-}
-
-.miniLabel .input {
-  --field--padding-top: calc(var(--space-base) + var(--space-smaller));
-  --field--padding-bottom: var(--space-small);
-}
-
-.textarea .input {
-  resize: vertical;
-}
-
-.small .input {
-  --field--padding-top: var(--space-small);
-  --field--padding-bottom: var(--space-small);
-}
-
-.select:not(.small) .input {
-  padding-top: 0;
-  padding-bottom: 0;
-}
-
-.miniLabel.select:not(.small) .input {
-  padding-top: var(--space-small);
-}
-
-/**
- * Label
- **/
-.label {
-  position: absolute;
-  top: 50%;
-  padding-left: var(--field--padding-left);
-  color: var(--field--placeholder-color);
-  line-height: calc(var(--base-unit) * 1.25);
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  pointer-events: none;
-  transform: translateY(-50%);
-  transition: all var(--timing-quick);
-}
-
-.miniLabel .label {
-  top: var(--space-smallest);
-  color: var(--color-text--secondary);
-  font-size: calc(var(--base-unit) * 0.875);
-  transform: translateY(0);
-}
-
-.textarea .label {
-  top: var(--field--padding-top);
-  transform: translateY(0);
-}
-
-.textarea.miniLabel .label {
-  top: var(--space-smallest);
-}
-
-.small.miniLabel .label {
-  display: none;
-}
-
-.large.miniLabel .label {
-  top: var(--space-small);
-}
-
-/**
- * Postfix
- **/
-.postfix {
-  position: absolute;
-  top: 50%;
-  right: var(--space-base);
-  z-index: calc(var(--field-base-elevation) - 1);
-  transform: translateY(-50%);
-}
