--- packages/components/src/FormField/FormFieldStyles.css.d.ts
@@ -1,16 +1,0 @@
-export const wrapper: string;
-export const textarea: string;
-export const invalid: string;
-export const disabled: string;
-export const small: string;
-export const large: string;
-export const inline: string;
-export const center: string;
-export const right: string;
-export const maxLength: string;
-export const inputWrapper: string;
-export const input: string;
-export const miniLabel: string;
-export const select: string;
-export const label: string;
-export const postfix: string;
